recent changes

+++ src/mincut_only.cpp
@@ -17,12 +17,12 @@ int MincutOnly::main() {
     int after_mincut_number_of_clusters = -2;
     int iter_count = 0;
 
-    std::map<std::string, int> original_to_new_id_map = ConstrainedClustering::GetOriginalToNewIdMap(&graph);
-    std::map<int, int> new_node_id_to_cluster_id_map = ConstrainedClustering::ReadCommunities(original_to_new_id_map, this->existing_clustering);
+    std::map<std::string, long> original_to_new_id_map = ConstrainedClustering::GetOriginalToNewIdMap(&graph);
+    std::map<long, long> new_node_id_to_cluster_id_map = ConstrainedClustering::ReadCommunities(original_to_new_id_map, this->existing_clustering);
     ConstrainedClustering::RemoveInterClusterEdges(&graph, new_node_id_to_cluster_id_map);
 
     /** SECTION Get Connected Components START **/
-    std::vector<std::vector<int>> connected_components_vector = ConstrainedClustering::GetConnectedComponents(&graph);
+    std::vector<std::vector<long>> connected_components_vector = ConstrainedClustering::GetConnectedComponents(&graph);
     /** SECTION Get Connected Components END **/
     if(this->connectedness_criterion == ConnectednessCriterion::Simple) {
         for(size_t i = 0; i < connected_components_vector.size(); i ++) {
